+++ pkg/providers/cps/provider.go
@@ -47,6 +47,7 @@ func Provider() *schema.Provider {
 		Schema: map[string]*schema.Schema{},
 		DataSourcesMap: map[string]*schema.Resource{
 			"akamai_cps_csr":         dataSourceCPSCSR(),
+			"akamai_cps_deployments": dataSourceDeployments(),
 			"akamai_cps_enrollment":  dataSourceCPSEnrollment(),
 			"akamai_cps_enrollments": dataSourceCPSEnrollments(),
 			"akamai_cps_warnings":    dataSourceCPSWarnings(),